switch to conan

+++ src/material.cpp
@@ -1,7 +1,7 @@
 #include "material.h"
 
 tiny::optional<std::pair<color3, ray>> lambertian::scatter(const ray& r_in, const hit_record& rec) const {
-    double random = random_double(0, 1);
+    float random = random_double(0, 1);
     if (random < -0.3) {
         return std::nullopt;
     }
@@ -23,11 +23,11 @@ tiny::optional<std::pair<color3, ray>> metal::scatter(const ray& r_in, const hit
 }
 
 tiny::optional<std::pair<color3, ray>> dielectric::scatter(const ray& r_in, const hit_record& rec) const {
-    double ri = rec.front_face ? (1.0/refraction_index) : refraction_index;
+    float ri = rec.front_face ? (1.0/refraction_index) : refraction_index;
 
     vec3 unit_direction = unit_vector(r_in.direction());
-    double cos_theta = std::fmin(dot(-unit_direction, rec.normal), 1.0);
-    double sin_theta = std::sqrt(1.0 - cos_theta*cos_theta);
+    float cos_theta = std::fmin(dot(-unit_direction, rec.normal), 1.0);
+    float sin_theta = std::sqrt(1.0 - cos_theta*cos_theta);
     vec3 direction;
 
     if (ri * sin_theta > 1.0 || reflectance(cos_theta, ri) > random_double()) { // cannot refract
